modifying projects and server cookies
modified demo1.jsx to be an iframe instead of an embed to the website. this now allows the user to press crtl + alt + i and open the inspect tool for specific projects on the landing page instead of just one project. added account creation api that returns cookies for account authorization. added different projects to the landing page made by the community.