rendering images part two
resolved issue with image buffer. figured out mongodb automatically converts the buffer into binary, which made it difficult to convert back. so the workaround is to convert the buffer to base64 and store that into mongodb, which is easier to handle than binary. now images can be hosted by the server even if the server fails or restarts.