figma light mode designs
created light mode designs for 3dinHTML on figma. light mode and dark mode will be determined mainly by the computer system settings but there's also a toggle on the landing page. i've always wanted to program a darkmode/lightmode site and this project is the perfect opportunity to do it.