server image url insertion and data
when the user uploads their code with images, the server will automatically host the image with its own url and that url will be inserted into the a-assets of the aframe code. this removes the problem of aframe projects only working due to local src files. when inserting code, adding a title, etc., going back and forth between the steps will now save when you return to that step.